+++ src/analyze/index.ts
@@ -8,6 +8,7 @@ import yaml from "yaml";
 import { Chunks } from "../utility/interfaces.js";
 import { OpenAPISpec } from "../utility/openapiGenerator.js";
 import initRules from "./helpers/initRules.js";
+import { EngineOutput, generateEngineOutput } from "./helpers/outputHelper.js";
 
 const availableTechs = {
     next: "Next.js",
@@ -36,14 +37,15 @@ const analyze = async (
     tech: "next",
     list: boolean,
     openapi: string,
-    validate: boolean
+    validate: boolean,
+    outputFile: string
 ) => {
     console.log(chalk.cyan(`[i] Loading analyze module...`));
 
     await initRules();
 
     // check if `-r` flag is there. If not, default to `~/.js-recon/rules`
-    if (!rulesPath) {
+    if (!rulesPath || rulesPath === "") {
         rulesPath = path.join(process.env.HOME, "/.js-recon/rules");
     }
 
@@ -120,13 +122,22 @@ const analyze = async (
     }
 
     // iterate over the ruleFiles
+    let ruleFindings: EngineOutput[] = [];
     for (const ruleFile of ruleFiles) {
         // load the rule
         const rule: Rule = yaml.parse(fs.readFileSync(ruleFile, "utf8"));
 
         // run the rule
-        await engine(rule, mappedJsonData, openapiData, tech);
+        const engineFindings: EngineOutput[] = await engine(rule, mappedJsonData, openapiData, tech);
+
+        // add findings to the global findings
+        if (engineFindings) {
+            ruleFindings.push(...engineFindings);
+        }
     }
+
+    // generate the engine output
+    generateEngineOutput(outputFile, ruleFindings);
 };
 
 export default analyze;