+++ CHANGELOG.md
@@ -5,11 +5,16 @@
 ### Added
 
 - Added `-k/--insecure` flag to disable SSL certificate verification
+- Added `json` output feature to analyze module
+- Added `analyze` module auto-run to `run` module
+- Added `report` module
 
 ### Changed
 
 - Updated the `postMessageFunctionResolve` step to resolve function expressions
 - Updated the `postMessageFunctionResolve` step to resolve arrow function expressions
+- Remove `--map-openapi` flag from run module (enabled by default)
+- Remove `--map-openapi-output` flag from run module
 
 ### Fixed